feat: base implementation
Some things are copied from existing lua plugin to not write boilerplate. bnf is a mixture of the plugin, luau and lua grammars that I'm trying to clean up at the moment.
It definitely lacks all the pin and fallback stuff, but I'm not sure it's all needed at the moment.