feat: variables reference contribution
It's a bit trickier than they say in example. You need to read docs. That's hard.

local_def_statement separation is not required in the end, but I guess it makes sense overall.
I still don't have a clear idea of mixin vs extend.
Having reference inside another node is not helpful, at least if it make reference optional. Because most of the methods don't allow you to return null.