chore: solve configuration cache problem
https://docs.gradle.org/8.6/userguide/configuration_cache.html#config_cache:requirements:disallowed_types
I asked Claude to solve the problem and it failed, so I did it myself through trial an errors. Actually, you need to read the docs closely. fs example is there, and they also say you can save reference to a project stuff on a task level. The rest two variables are available from the task itself.