fix: type casts in the binary expressions parsing
Previous order was stupid because `simple_exp` being a part of `as_exp` will be parsed first and then try to recover or just fail to parse.
But just switching place of `as_exp` and `simple_exp` didn't help. I tried to debug but get little info from that apart from the parsing is weird. So I tried to put the most complex expressions first and it seems to work. I'll see how it goes.