feat: reverse fake table field
It's generating methods that are annotated incorrectly, because there can be no id. Just implementing handcrafting TableField interface doesn't generate `fieldList` property on LuauTableConstructor. So it has to be added manually. Well, I'm too lazy now, and I looked at Haxe, and they keep the alternative of the table field in PSI, so it should be not that bad. Reconsider later