feat: improve error recovery
I thought about pinning an index_or_call_expr
```
//  There is a hard choice here. This expr is the last thing that can be tried for a statement,
//  so we can pin it to the first element.
//  Because if you started writing an id where statement is expected, most probably you wanted
//  the function call and didn't finish or made a mistake. But pinning this would break the value expression above.
//  Because it will always pin to the first branch and won't try the rest.
//  There is one more thing that can go wrong, you can write 1 + 1 or just any other non simple reference thing, like "".
//  This can't be considered as an index access, but I still want this to be eaten by the block and I don't know how.
```
But then I took a look at the rust bnf and use an interesting technic by pinning the statement unless it's the end of the block `!'}'`. So it does what I wanted it considers everything a statement until the end, so block includes all the errors.

I imagine there is no way to start statement parsing after the `,`, so I dropped it from the recover rule. Same goes for unary operator, they can only start expressions, and these expressions are not allowed as statements. Same for string, number, nil and booleans