feat: improve error handling in stylua autoformat
I drop a bunch of old code. It was mostly copied from the prettier plugin, and I didn't know whether I needed it.
DocumentUpdatingActionOnSave is the recommended way to update files and it's example is easier than the one I was using.
Also, I found in Rust plugin that I can pass input to the Stylua after creation.