feat: enable back lsp restart
Now I correctly compare the stuff so the lsp is restarted straight away when I disable it.
I moved Event and notify implementations to the configurable because it's now the place where all the changes are applied. So it doesn't make much sense to keep in the state, and they all are just compile-time things and a static variable.