fix: join reference and psi element implementation
Well, I spend a lot of time looking into different implementations and ... they are all different. Given that I don't have multiple references at the moment and that I want to use the builtin ResolveCache, having stable reference is important. So I went the Dart/Haxe path, where they join PsiElement and PsiReference and I didn't fine a way use PsiReferenceBase, because, even with delegate it should be one "level" above the element and take it as the parameter. I was wondering if I can just store the reference in the field, but I'm not sure how safe is this.