feat: lsp managment
Well, I suffered a lot along the way. Definitely, this settings version is not final. And I have at least to give an ability to delete older LSP versions. The main takeaway is that finding how to implement comboboxes is hard. Trying to download stuff on applying is tricky, because it's done not in a coroutine, so you have to spawn one, but then you can't block the saving if download failed.
Notifications are hard as well; I still don't know how to show action as a button not as a link.
How to make reactive bindings to update settings is still a mystery as well. LLM always wants a PropertyGraph, but I found no usages of it.
Default settings are a lame way to ask less configuration. Because half the settings are easily application wide and half are project wide. And I haven't found a good example of how it can be done with builtin ways, neither I can think of a good separation to provide two separate project wide and app wide settings.