feat: check lsp on lsp server start
Generation of cli seems to be done only when the server is really going to start. It's not even rerun if trying to run it the first time failed. Should be good for my purposes and maybe even closer than the previous one. Given that people may have multilanguage projects, and being notified that some lsp is not there is not productive.
The downside is that restarting the service will call the check again. Shouldn't be that bad though. 1) I don't think people restart LSP a lot 2) network request for LSP version is cached if they do that a lot 3) The rest (mostly checking files presence) are fast.