fix: incorrect initialization of LSP with auto configuration when you open the new project
The initialization is still incorrect, since I start the LSP synchronously, I can't check the FS for lsps and update the settings. But I now restart the lsp once the settings are correctly set.
I also simplified the check for lsp configuration. I don't care much about lspVersion being null, given this state is not possible to achieve through UI, and it can only be the result of some corrupted state the check is not useful.
I also switched to a simpler check instead of getting the full configuration in the `fileOpened` since I don't need a full config there, and I can get it only once when actually starting the LSP. The problem is that the checks are related but written in two places, so they can diverge in time. I have to address this later.