feat: separate roblox api definitions from LSP and concentrate LSP restart into the service
The first is done by example of the LSP, and I think it is done to decouple API updates from the LSP release cycle that makes sense. I used the same pattern of checking definitions on the start of lsp once a day and downloading them if they are missing or stale
The second has been done along the way, since I need to react to the definitions update with the lsp restart and I tired of using restart in different places and it's better to be a centralized service. Now I can also rerun lsps in the open projects when the lsp is updated.