feat: minimal configuration of lsp
Support lsp workspace configuration for the basic options I already have.
Add platform type since it affects the lsp require resolving.
Fix restart service but converting it to the declaratively registered listeners. I always forget that service is not created if you do not get its instance somewhere. Adding one more postStartupActivity is lame, so I went the preffered way by adding listeners in the xml config.
Along the way fix missing fields in the default settings.
I tried to gather lsp settings in a nested data class, but haven't found a lot of benefits, but there are a couple downsides - I have to migrate the settings plus it's hard to get to internal properties in the Change event listeners.