fix: restarting lsp
restart is used also to stop the running lsp. So the check if we need to start an LSP is bundled into LSP provider, there is no need for now to optimize it and ask before restart. Otherwise, I need to write the checks for whether the lsp is currently running and stop it.