feat: make lsp more robust and cleanup old parts
I wanted to download chosen LSP only when you apply settings before, but I went with downloading it upon selection, because downloading them on application is a bit more tricky (there is no way to detect the save button click, so the download has to happen in the background, and I wanted it to be sync to show errors straight away). Hence, all isLspVersionModified similar checks are removed.
I also no longer store lsp version in the lspVersion field if the latest is used, so the 1634 lines are dropped because the field will be updated anyway when the correct version is chosen.