fix: usage of internal API
I still don't quite get how to correctly handle these things. Most of the plugins using this are internal and have their own modules defined which have the files to exclude defined already.
I tried to find them using FilenameIndex.processFilesByNames and got a warning that you cannot use it until the project is loaded. Is there a way to know if the project is loaded? No idea